# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "cristinaperezboix.alu@iespacomolla.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/characters/{id}/edit": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "Muestra el formulario para editar un personaje específico",
                "description": "Muestra el formulario para editar un personaje específico.",
                "operationId": "050f5684b70a825363ea1a1f8af29b0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del personaje",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vista del formulario para editar el personaje"
                    },
                    "404": {
                        "description": "Character not found"
                    }
                }
            }
        },
        "/characters/{id}": {
            "put": {
                "tags": [
                    "Characters"
                ],
                "summary": "Actualiza un personaje específico",
                "description": "Actualiza un personaje específico.",
                "operationId": "821d2b3fd3bc58de60d21da164a85fcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del personaje",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "El nombre del personaje",
                                        "type": "string"
                                    },
                                    "class": {
                                        "description": "La clase del personaje",
                                        "type": "string"
                                    },
                                    "race": {
                                        "description": "La raza del personaje",
                                        "type": "string"
                                    },
                                    "faction": {
                                        "description": "La facción del personaje",
                                        "type": "string"
                                    },
                                    "background": {
                                        "description": "El trasfondo del personaje",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Character updated successfully"
                    },
                    "404": {
                        "description": "Character not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Characters"
                ],
                "summary": "Elimina el personaje especificado",
                "description": "Elimina un personaje específico.",
                "operationId": "a157fcbefc13849cfc3bfc85f0daaa1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del personaje",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful character delete"
                    },
                    "403": {
                        "description": "You have no permission"
                    },
                    "404": {
                        "description": "Character not found"
                    }
                }
            }
        },
...
    },
...
}